app/requirements.txt
fastapi==0.82.0
uvicorn==0.18.3
pydantic~=1.10.2
bcrypt==4.0.1
python-multipart==0.0.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
mysql-connector-python
